backend-api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
docstring_parser==0.16
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.85.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.25.1
google-crc32c==1.7.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
